requirements.txt
fastapi
uvicorn
easyocr
opencv-python
numpy
pdfplumber
pillow
python-docx
pymupdf
reportlab
nltk
python-multipart